pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.poetry]
authors = ["JimZer <jimi.vaubien@protonmail.com>"]
description = "A CLI tool to help lazy coders get the work done with AI (commit messages, pull requests ...)"
homepage = "https://gprithub.com/bitswired/lazycodr"
name = "lazycodr"
packages = [{include = "lazycodr", from = "src"}]
readme = "README.md"
repository = "https://github.com/bitswired/lazycodr"
version = "0.1.2"

[tool.poetry.dependencies]
httpx = "^0.24.0"
jinja2 = "^3.1.2"
langchain = "^0.0.325"
openai = "^0.28.1"
pathspec = "^0.11.2"
pyfiglet = "^1.0.2"
pygithub = "^2.1.1"
python = "^3.10"
rich = "^13.6.0"
tiktoken = "^0.4.0"
tqdm = "^4.66.1"
typer = {extras = ["all"], version = "^0.9.0"}

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"
pytest = "^7.4.3"
ruff = "^0.1.3"

[tool.poetry.scripts]
lazycodr = "lazycodr.cli:main"

[tool.ruff]
ignore = ['D100', 'D103', 'D104', 'D107', 'ANN', 'BLE001', 'FA102']
select = ['ALL']
# Note: Ruff supports a top-level `src` option in lieu of isort's `src_paths` setting.
src = ["src", "tests"]

[tool.ruff.flake8-bugbear]
extend-immutable-calls = ["typer.Option"]

[tool.ruff.isort]
known-first-party = ["lazycodr"]

[tool.ruff.per-file-ignores]
"tests/**/*" = ['S101']
